requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.24.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.6
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.58.4
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
keyboard==0.13.5
kiwisolver==1.4.8
matplotlib==3.10.3
microsoft-kiota-abstractions==1.9.5
microsoft-kiota-authentication-azure==1.9.5
microsoft-kiota-http==1.9.5
microsoft-kiota-serialization-form==1.9.5
microsoft-kiota-serialization-json==1.9.5
microsoft-kiota-serialization-multipart==1.9.5
microsoft-kiota-serialization-text==1.9.5
msal==1.33.0
msal-extensions==1.3.1
msgraph-core==1.3.5
multidict==6.6.4
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.2.1
plyer==2.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
std-uritemplate==2.0.5
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
zipp==3.23.0
